requirements.txt
lxml
pydantic
python-dotenv
pyyaml
rich
networkx
requests

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0

# LLM / RAG (choose one runtime later)
langchain
chromadb
tiktoken

# Web application
flask>=3.0.0

# Documentation formats
python-docx>=1.0.0
markdown
weasyprint
